Move auth from lib to server folder
